Improved documentation in man pages.
Ticket: UD-1248
Reviewed By: nobody
Testing Done:

Tried to be clearer about default ports.   Added a comment in the
plugin manpage that debug can be set in the config file as well.

There was a complaint that it wasn't clear on how to set the various
options for the pam plugin module, so I tried to make that cleaer
as well.

Also fixed some spelling errors in the man page, and remove the
non-debian config file default.