Fixed problem with 2nd config init, when no servers in config file
Ticket: CM-20454
Reviewed By: nobody

My code to avoid redoing all the config didn't work right when
re-entered with no server listed in the config file.  The result
was I'd return an error the first time, and success the 2nd-Nth
times, and then later code would try to dereference the NULL pointer
server list, and segv in login or sshd, etc.

Redid the logic in initialize() to fix that.